feat: adiciona tratamento de exceção global
Adiciona tratamento de exceção em nível Global com ControllerAdvice e CustomError e adiciona tratamento de exceções nos métodos de update e delete, além de adicionar nova exceção personalizada para falha de integridade relacional no banco.

+++ java-spring-professional/apirest-layers-crud-exceptions-validations/challenge/src/main/java/com/raphaelsena/crud_clientes/services/ClientService.java
@@ -1,11 +1,16 @@
 package com.raphaelsena.crud_clientes.services;
 
+import com.fasterxml.jackson.databind.exc.InvalidDefinitionException;
 import com.raphaelsena.crud_clientes.models.Client;
 import com.raphaelsena.crud_clientes.models.dtos.ClientDTO;
 import com.raphaelsena.crud_clientes.repositories.ClientRepository;
+import com.raphaelsena.crud_clientes.services.exceptions.DatabaseException;
 import com.raphaelsena.crud_clientes.services.exceptions.ObjectNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -29,8 +34,12 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO obj) {
-        Client client = clientRepository.getReferenceById(id);
-        return fromDTO(obj, client);
+        try {
+            Client client = clientRepository.getReferenceById(id);
+            return fromDTO(obj, client);
+        } catch (EntityNotFoundException e) {
+            throw new ObjectNotFoundException("Objeto não encontrado.");
+        }
     }
 
     private ClientDTO fromDTO(ClientDTO obj, Client client) {
@@ -45,8 +54,16 @@ public class ClientService {
         return new ClientDTO(client);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ObjectNotFoundException("Objeto não encontrado");
+        }
+
+        try {
+            clientRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial. Há entidades relacionadas a este cliente.");
+        }
     }
 }
